+++ src/action-sheet/index.tsx
@@ -1,6 +1,6 @@
 import getCurrentPage from '../get-current-page';
-import React, { useEffect, useRef, useState } from 'react';
-import Sheet, { SheetInstance } from '../sheet';
+import React, { useRef, useState } from 'react';
+import { SheetWrap, SheetWrapInstance, SheetWrapData } from '../sheet';
 import { View } from 'remax/one';
 import styles from './index.module.less';
 import Space from '../space';
@@ -9,59 +9,54 @@ export interface ShowOptions {
   items: { label: React.ReactNode; value: string | number }[];
 }
 
-const data: {
-  [route: string]: (
-    options: ShowOptions,
-  ) => Promise<ShowOptions['items'][number]>;
-} = {};
+const data: SheetWrapData = {};
 
 const ActionSheet = () => {
-  const page = getCurrentPage();
-  const [items, setItems] = useState<ShowOptions['items']>([]);
-  const ref = useRef<SheetInstance>(null);
-  const promiseRef = useRef<any>({ resolve: () => {}, reject: () => {} });
-  useEffect(() => {
-    data[page] = ({ items }) =>
-      new Promise((resolve, reject) => {
-        setItems(items);
-        ref.current?.setVisible(true);
-        promiseRef.current = { reject, resolve };
-      });
-  }, [page]);
+  const [options, setOptions] = useState<{ items: ShowOptions['items'] }>({
+    items: [],
+  });
+  const ref = useRef<SheetWrapInstance>(null);
   return (
-    <Sheet ref={ref}>
+    <SheetWrap ref={ref} setOptions={setOptions} data={data}>
       <Space vertical style={{ width: '100vw' }}>
-        {items.map(({ label, value }, index) => (
+        {options.items.map(({ label, value }, index) => (
           <React.Fragment key={index}>
             <View
               className={styles.item}
               onTap={() => {
-                ref.current?.setVisible(false);
-                promiseRef.current.resolve({ label, value });
+                ref.current?.sheetRef?.setVisible(false);
+                ref.current?.promiseRef.resolve({ label, value });
               }}
             >
               {label}
             </View>
-            {index !== items.length - 1 && <View className={styles.line} />}
+            {index !== options.items.length - 1 && (
+              <View className={styles.line} />
+            )}
           </React.Fragment>
         ))}
         <View
           className={styles.cancel}
           onTap={() => {
-            ref.current?.setVisible(false);
-            promiseRef.current.reject();
+            ref.current?.sheetRef?.setVisible(false);
+            ref.current?.promiseRef.reject();
           }}
         >
           取消
         </View>
       </Space>
-    </Sheet>
+    </SheetWrap>
   );
 };
 
 ActionSheet.show = (options: ShowOptions) => {
   const page = getCurrentPage();
-  return data[page](options);
+  return data[page].fn(options);
+};
+
+ActionSheet.hide = () => {
+  const page = getCurrentPage();
+  return data[page].setVisible?.(false);
 };
 
 export default ActionSheet;