+++ src/picker/index.wechat.tsx
@@ -11,56 +11,17 @@ import { Props, useProps } from './common';
 import { View } from 'remax/one';
 import dayjs from 'dayjs';
 
-const dateTimeData = new Array(1).fill(0).map((_, i) => {
-  const y = dayjs()
-    .add(i, 'y')
-    .year();
-  return {
-    value: y,
-    label: y + '年',
-    children: new Array(12 - dayjs().get('month')).fill(0).map((_, i) => {
-      const m = i + 1 + dayjs().get('month');
-      return {
-        label: m + '月',
-        children: new Array(
-          dayjs(`${y}-${m}`)
-            .endOf('month')
-            .date(),
-        )
-          .fill(0)
-          .map((_, i) => {
-            const d = i + 1;
-            return {
-              label: d + '日',
-              children: new Array(24).fill(0).map((_, i) => {
-                const h = i + 1;
-                return {
-                  label: (h < 10 ? '0' + h : h) + '时',
-                  children: new Array(60).fill(0).map((_, i) => {
-                    return {
-                      label: (i < 10 ? '0' + i : i) + '分',
-                    };
-                  }),
-                };
-              }),
-            };
-          }),
-      };
-    }),
-  };
-});
-
 export default (props: Props) => {
   const {
     data = useMemo(() => [], []),
     value,
     onChange,
     children,
-    cols,
     start,
     end,
     childrenCls,
     mode = 'multiSelector',
+    cols = mode === 'datetime' ? 5 : undefined,
     ...newProps
   } = useProps(props);
   const getData = useCallback(
@@ -69,6 +30,84 @@ export default (props: Props) => {
     [],
   );
   const [columnIndex, setColumnIndex] = useState<number[]>([]);
+  const isDatetime = mode === 'datetime';
+  const [
+    yIndex = 0,
+    mIndex = 0,
+    dIndex = 0,
+    hIndex = 0,
+    mmIndex = 0,
+  ] = columnIndex;
+  const minuteData = useMemo(() => {
+    if (!isDatetime) return [];
+    const ys = new Array(dayjs(end).diff(start, 'y') + 1).fill(0).map((_, i) =>
+      dayjs(start)
+        .add(i, 'y')
+        .format('YYYY年'),
+    );
+    const getArr = (
+      parent: string[],
+      parentIndex: number,
+      unit: dayjs.UnitType,
+      cnUnit: string,
+      limit: number,
+      isStart: boolean,
+      isEnd: boolean,
+    ) => {
+      const startIndex = dayjs(start).get(unit);
+      const endIndex = dayjs(end).get(unit);
+      const length = isStart
+        ? limit - startIndex
+        : isEnd
+        ? endIndex + 1
+        : limit;
+      return new Array(
+        unit === 'date'
+          ? isEnd
+            ? length - 1
+            : isStart
+            ? length + 1
+            : length
+          : length,
+      )
+        .fill(0)
+        .map((_, i) => {
+          let value: any =
+            (isStart ? startIndex + i : i) +
+            (['hours', 'minutes'].includes(unit) || (unit === 'date' && isStart)
+              ? 0
+              : 1);
+          value = value < 10 ? `0${value}` : value;
+          return `${value}${cnUnit}`;
+        });
+    };
+    const isStartY = yIndex === 0;
+    const isEndY = yIndex === ys.length - 1;
+    const ms = getArr(ys, yIndex, 'months', '月', 12, isStartY, isEndY);
+    const isStartM = isStartY && mIndex === 0;
+    const isEndM = isEndY && mIndex === ms.length - 1;
+    const ds = getArr(
+      ms,
+      mIndex,
+      'date',
+      '日',
+      dayjs(
+        `${ys[yIndex]}${ms[mIndex]}`.replace(/年/, '-').replace(/月/, ''),
+        'YYYY-M',
+      )
+        .endOf('months')
+        .date(),
+      isStartM,
+      isEndM,
+    );
+    const isStartH = isStartM && dIndex === 0;
+    const isEndH = isEndM && dIndex === ds.length - 1;
+    const hs = getArr(ds, hIndex, 'hours', '时', 24, isStartH, isEndH);
+    const isStartMm = isStartH && hIndex === 0;
+    const isEndMm = isEndH && hIndex === hs.length - 1;
+    const mms = getArr(ds, mmIndex, 'minutes', '分', 60, isStartMm, isEndMm);
+    return [ys, ms, ds, hs, mms];
+  }, [dIndex, end, hIndex, isDatetime, mIndex, mmIndex, start, yIndex]);
   const range = useMemo(() => {
     const range: (PickerData[] | PickerData[][])[] = [];
     new Array(cols).fill(0).forEach((_, index) => {
@@ -76,11 +115,10 @@ export default (props: Props) => {
         index === 0 ? data : getData(range[index - 1], columnIndex[index - 1]),
       );
     });
-    return range;
-  }, [cols, columnIndex, data, getData]);
+    return isDatetime ? minuteData : range;
+  }, [cols, columnIndex, data, getData, isDatetime, minuteData]);
   const rangeRef = useRef(range);
   rangeRef.current = range;
-
   useEffect(() => {
     const numbers: number[] = [];
     /** 默认值 */
@@ -88,56 +126,36 @@ export default (props: Props) => {
       setColumnIndex(new Array(cols).fill(0));
       return;
     }
-    if (cols === 1 && ['string', 'number'].includes(typeof value)) {
+    if (isDatetime) {
+      let newValue: any = value;
+      newValue = newValue
+        .split(/-|\s|:/)
+        .map(
+          (str: any, i: any) => `${str}${['年', '月', '日', '时', '分'][i]}`,
+        );
+      newValue.forEach((v: any, i: any) => {
+        numbers[i] = rangeRef.current[i].findIndex((value: any) => v === value);
+        numbers[i] = numbers[i] === -1 ? 0 : numbers[i];
+      });
+    } else if (cols === 1 && ['string', 'number'].includes(typeof value)) {
       numbers[0] = rangeRef.current[0]
         .flat()
-        .findIndex(({ value: v }) => v === value);
+        .findIndex(({ value: v }: any) => v === value);
     } else if (Array.isArray(value)) {
       new Array(cols).fill(0).forEach((_, index) => {
         numbers[index] = value?.[index]
           ? rangeRef.current[index]
               .flat()
-              .findIndex(({ value: v }) => value?.[index] === v)
+              .findIndex(({ value: v }: any) => value?.[index] === v)
           : 0;
       });
     }
 
     setColumnIndex(numbers);
-  }, [cols, value]);
-
+  }, [cols, isDatetime, value]);
   const isDateOrTimeMode = useMemo(() => {
     return ['date', 'time'].includes(mode);
   }, [mode]);
-  const [, mIndex = 0, dIndex = 0, hIndex = 0] = columnIndex;
-  const isDatetime = mode === 'datetime';
-  const minuteData = useMemo(() => {
-    if (!isDatetime) return [];
-    const y = dateTimeData[0].label;
-    const m = dateTimeData[0].children.map(({ label }) => label);
-    const d = dateTimeData[0].children[mIndex].children
-      .filter(
-        (_, i) =>
-          m[mIndex] !== dayjs().get('month') + 1 + '月' ||
-          i + 1 >= dayjs().get('date'),
-      )
-      .map(({ label }) => label);
-    const h = dateTimeData[0].children[0].children[0].children
-      .filter(
-        (_, i) =>
-          d[dIndex] !== dayjs().get('date') + '日' || i + 1 >= dayjs().get('h'),
-      )
-      .map(({ label }) => label);
-    const mm = dateTimeData[0].children[0].children[0].children[0].children
-      .filter(
-        (_, i) =>
-          h[hIndex] !== dayjs().get('hour') + '时' ||
-          i + 1 >= dayjs().get('minute'),
-      )
-      .map(({ label }) => label);
-    return [[y], m, d, h, mm];
-  }, [dIndex, hIndex, isDatetime, mIndex]);
-  console.log(minuteData);
-
   return (
     <Picker
       {...newProps}
@@ -145,10 +163,15 @@ export default (props: Props) => {
       mode={isDatetime ? 'multiSelector' : (mode as any)}
       start={start}
       end={end}
-      range={isDatetime ? minuteData : range}
+      range={range}
       value={!isDateOrTimeMode ? columnIndex : value}
       onColumnChange={({ detail: { column, value } }) => {
         columnIndex[column] = value;
+        columnIndex.forEach((_, i) => {
+          if (i > column) {
+            columnIndex[i] = 0;
+          }
+        });
         setColumnIndex([...columnIndex]);
       }}
       onChange={({ detail: { value } }) => {