+++ src/set-navigation-bar-color/index.web.ts
@@ -1,0 +1,1 @@
+export default () => console.log('set-navigation-bar-color');